Add city check.

+++ app/app/src/main/java/com/weather/app/entity/detail/Main.kt
@@ -20,9 +20,9 @@ data class Main (
 	@SerializedName("feels_like") val feels_like : Double,
 	@SerializedName("temp_min") val temp_min : Double,
 	@SerializedName("temp_max") val temp_max : Double,
-	@SerializedName("pressure") val pressure : Int,
-	@SerializedName("sea_level") val sea_level : Int,
-	@SerializedName("grnd_level") val grnd_level : Int,
-	@SerializedName("humidity") val humidity : Int,
+	@SerializedName("pressure") val pressure : Double,
+	@SerializedName("sea_level") val sea_level : Double,
+	@SerializedName("grnd_level") val grnd_level : Double,
+	@SerializedName("humidity") val humidity : Double,
 	@SerializedName("temp_kf") val temp_kf : Double
 )
